Multiple Target support (x86_64, armv7)
+ extra packages like openssh (by default on ARM)
+ review network hook (support 3.10 x86_64 kernel)
+ cower for pacaur (binary not available for x86_64)

+++ builder/sudo-workaround.sh
@@ -26,9 +26,11 @@ EOF
 }
 
 function uninstall() {
-	rm /usr/bin/sudo
-	mv /usr/bin/sudo.workaround /usr/bin/sudo
-	rm /usr/bin/qemu-arm-staticS
+	if [ -f /usr/bin/sudo.workaround ]; then 
+		rm /usr/bin/sudo
+		mv /usr/bin/sudo.workaround /usr/bin/sudo
+		rm /usr/bin/qemu-arm-staticS
+	fi
 }
 
 # Needs sudo installed
